+++ WeatherApiMvcApp/Models/ApiModels/Sys.cs
@@ -7,11 +7,11 @@ namespace WeatherApiMvcApp.Models.ApiModels
 {
     public class Sys
     {
-        public int type { get; set; }
-        public int id { get; set; }
+        public double type { get; set; }
+        public double id { get; set; }
         public double message { get; set; }
         public string country { get; set; }
-        public int sunrise { get; set; }
-        public int sunset { get; set; }
+        public double sunrise { get; set; }
+        public double sunset { get; set; }
     }
 }
